Refactor: Extract UI layout logic to layout.js
