+++ src/index.js
@@ -1,19 +1,10 @@
 import './style.css';
-import Icon from './logo.png';
 import * as sim from './core/simulation';
 import { appState, fibonacciCalendarMappings, tshirtMappings } from './core/state';
-import saveSvgAsImage from './utils/export-utils';
-import { applyGraphSettings, resetGraphSettings, createAdvancedSettings } from './ui/graph-settings';
-import {
-  createModeSelector,
-  createFileLoader,
-  createDataEntrySection,
-} from './data-input';
+import { applyGraphSettings, resetGraphSettings } from './ui/graph-settings';
 import { gatherRawTaskData, normalizeTaskData } from './ui/task-table';
+import { setupUi } from './ui/layout';
 import {
-  createDivWithIdAndClasses,
-  createTextElement,
-  createLabeledInput,
   showError,
 } from './utils/dom-helpers';
 
@@ -252,209 +243,6 @@ async function startSimulation(event) {
   }
 }
 
-/**
- * Creates the project logo image element.
- * @returns {HTMLImageElement} Logo image
- */
-function createLogoElement() {
-  const simIcon = document.createElement('img');
-  simIcon.src = Icon;
-  simIcon.alt = 'Project Estimate Simulator icon';
-  simIcon.width = 100;
-  simIcon.height = 100;
-  simIcon.classList.add('project-icon');
-  return simIcon;
-}
-
-/**
- * Creates the page header with GitHub ribbon.
- * @returns {HTMLElement} Header section
- */
-function createHeader() {
-  const headerDiv = document.createElement('div');
-  headerDiv.classList.add('page-header', 'section');
-
-  // Fork Me ribbon
-  const githubRibbon = createDivWithIdAndClasses('forkOnGithub', ['github-ribbon']);
-  const githubLink = document.createElement('a');
-  githubLink.href = 'https://github.com/acrosman/simple-project-estimates';
-  githubLink.textContent = 'Fork me on GitHub';
-  githubRibbon.appendChild(githubLink);
-  headerDiv.appendChild(githubRibbon);
-
-  return headerDiv;
-}
-
-/**
- * Creates the simulation control panel and results display.
- * @returns {HTMLElement} Simulation panel section
- */
-function createSimulationPanel() {
-  const simWrapper = createDivWithIdAndClasses('simulationAreaWrapper', ['section', 'container']);
-  const simHeader = createTextElement('h2', 'Simulator', ['header', 'simulation']);
-
-  const simControls = createDivWithIdAndClasses('simulatorControlsWrapper', ['section', 'controls-simulation']);
-
-  const simCountFldAttr = {
-    type: 'number',
-    min: '1000',
-    max: '9999999',
-    step: '1000',
-    id: 'simulationPasses',
-    value: '100000',
-    name: 'Simulation Passes',
-  };
-  const simCountCtl = createLabeledInput('Number of times to run the simulation:', simCountFldAttr, true);
-
-  const simLimitFldAttr = {
-    type: 'checkbox',
-    value: '1',
-    id: 'LimitGraph',
-  };
-  const simLimitCtl = createLabeledInput('Limit graph outliers', simLimitFldAttr, false);
-
-  const simRun = document.createElement('input');
-  simRun.type = 'button';
-  simRun.id = 'startSimulationButton';
-  simRun.value = 'Run Simulation';
-  simRun.addEventListener('click', startSimulation);
-
-  simControls.appendChild(simCountCtl);
-  simControls.appendChild(simLimitCtl);
-  simControls.appendChild(simRun);
-
-  // Simulation Time Results elements
-  const simResultWrapper = createDivWithIdAndClasses('simulationResultsWrapper', ['section', 'wrap-simulation-results']);
-  simResultWrapper.appendChild(createDivWithIdAndClasses('simulationRunningTime', ['simulation-result', 'text']));
-  const simTimeResultWrapper = createDivWithIdAndClasses('simulationTimeResultsWrapper', ['section', 'wrap-simulation-time-results']);
-  const timeHeader = createTextElement('h3', 'Time Estimates', ['result-display', 'time-info']);
-  timeHeader.id = 'timeEstimateHeader';
-  timeHeader.style.display = 'none';
-  simTimeResultWrapper.appendChild(timeHeader);
-  simTimeResultWrapper.appendChild(createDivWithIdAndClasses('simulationTimeMedian', ['simulation-result', 'time-info', 'text']));
-  simTimeResultWrapper.appendChild(createDivWithIdAndClasses('simulationTimeStandRange', ['simulation-result', 'time-info', 'text']));
-  simTimeResultWrapper.appendChild(createDivWithIdAndClasses('simulationTimeMax', ['simulation-result', 'time-info', 'text']));
-  simTimeResultWrapper.appendChild(createDivWithIdAndClasses('simulationTimeMin', ['simulation-result', 'time-info', 'text']));
-  simTimeResultWrapper.appendChild(createDivWithIdAndClasses('simulationTimeStandDev', ['simulation-result', 'time-info', 'text']));
-  simTimeResultWrapper.appendChild(createDivWithIdAndClasses('timeHistoGram', ['simulation-result', 'time-info', 'graph']));
-
-  // Add save buttons for time histogram
-  const timeSaveButtonsDiv = document.createElement('div');
-  timeSaveButtonsDiv.id = 'timeSaveButtons';
-  timeSaveButtonsDiv.classList.add('save-buttons', 'no-print');
-  timeSaveButtonsDiv.style.display = 'none';
-
-  const saveTimePng = document.createElement('input');
-  Object.assign(saveTimePng, {
-    type: 'button',
-    value: 'Save Time Graph as PNG',
-    id: 'saveTimePngBtn',
-  });
-  saveTimePng.addEventListener('click', () => saveSvgAsImage('timeHistoGram', 'time-estimates', 'png'));
-
-  const saveTimeJpeg = document.createElement('input');
-  Object.assign(saveTimeJpeg, {
-    type: 'button',
-    value: 'Save Time Graph as JPEG',
-    id: 'saveTimeJpegBtn',
-  });
-  saveTimeJpeg.addEventListener('click', () => saveSvgAsImage('timeHistoGram', 'time-estimates', 'jpeg'));
-
-  timeSaveButtonsDiv.appendChild(saveTimePng);
-  timeSaveButtonsDiv.appendChild(saveTimeJpeg);
-  simTimeResultWrapper.appendChild(timeSaveButtonsDiv);
-
-  simResultWrapper.appendChild(simTimeResultWrapper);
-
-  // Simulation Cost Results elements
-  const simCostResultWrapper = createDivWithIdAndClasses('simulationCostResultsWrapper', ['section', 'wrap-simulation-cost-results']);
-  const costHeader = createTextElement('h3', 'Cost Estimates', ['result-display', 'cost-info']);
-  costHeader.id = 'costEstimateHeader';
-  costHeader.style.display = 'none';
-  simCostResultWrapper.appendChild(costHeader);
-  simCostResultWrapper.appendChild(createDivWithIdAndClasses('simulationCostMedian', ['simulation-result', 'cost-info', 'text']));
-  simCostResultWrapper.appendChild(createDivWithIdAndClasses('simulationCostStandRange', ['simulation-result', 'cost-info', 'text']));
-  simCostResultWrapper.appendChild(createDivWithIdAndClasses('simulationCostMax', ['simulation-result', 'cost-info', 'text']));
-  simCostResultWrapper.appendChild(createDivWithIdAndClasses('simulationCostMin', ['simulation-result', 'cost-info', 'text']));
-  simCostResultWrapper.appendChild(createDivWithIdAndClasses('simulationCostStandDev', ['simulation-result', 'cost-info', 'text']));
-  simCostResultWrapper.appendChild(createDivWithIdAndClasses('costHistoGram', ['simulation-result', 'cost-info', 'graph']));
-
-  // Add save buttons for cost histogram
-  const costSaveButtonsDiv = document.createElement('div');
-  costSaveButtonsDiv.id = 'costSaveButtons';
-  costSaveButtonsDiv.classList.add('save-buttons', 'no-print');
-  costSaveButtonsDiv.style.display = 'none';
-
-  const saveCostPng = document.createElement('input');
-  Object.assign(saveCostPng, {
-    type: 'button',
-    value: 'Save Cost Graph as PNG',
-    id: 'saveCostPngBtn',
-  });
-  saveCostPng.addEventListener('click', () => saveSvgAsImage('costHistoGram', 'cost-estimates', 'png'));
-
-  const saveCostJpeg = document.createElement('input');
-  Object.assign(saveCostJpeg, {
-    type: 'button',
-    value: 'Save Cost Graph as JPEG',
-    id: 'saveCostJpegBtn',
-  });
-  saveCostJpeg.addEventListener('click', () => saveSvgAsImage('costHistoGram', 'cost-estimates', 'jpeg'));
-
-  costSaveButtonsDiv.appendChild(saveCostPng);
-  costSaveButtonsDiv.appendChild(saveCostJpeg);
-  simCostResultWrapper.appendChild(costSaveButtonsDiv);
-
-  simResultWrapper.appendChild(simCostResultWrapper);
-
-  // Add simulator elements to wrapper.
-  simWrapper.appendChild(simHeader);
-  simWrapper.appendChild(simControls);
-  simWrapper.appendChild(createAdvancedSettings());
-  simWrapper.appendChild(simResultWrapper);
-
-  return simWrapper;
-}
-
-/**
- * Setup the Main application UI
- * @returns HTMLElement
- */
-function setupUi() {
-  // === Main page structures ===
-  const mainElement = document.createElement('div');
-
-  // Setup data entry section wrapper
-  const dataWrapper = document.createElement('div');
-  dataWrapper.classList.add('section');
-  dataWrapper.id = 'dataAreaWrapper';
-
-  // Build UI sections
-  const header = createHeader();
-  const modeSelector = createModeSelector();
-  const fileLoader = createFileLoader();
-  const dataEntry = createDataEntrySection();
-  const simulationPanel = createSimulationPanel();
-
-  // Create side-by-side wrapper: mode selector on the left, logo on the right
-  const modeSelectorWithLogo = document.createElement('div');
-  modeSelectorWithLogo.classList.add('mode-selector-logo-wrapper');
-  modeSelectorWithLogo.appendChild(modeSelector);
-  modeSelectorWithLogo.appendChild(createLogoElement());
-
-  // Assemble data area
-  dataWrapper.appendChild(modeSelectorWithLogo);
-  dataWrapper.appendChild(fileLoader);
-  dataWrapper.appendChild(dataEntry);
-
-  // Add all elements to the main application wrapper.
-  mainElement.appendChild(header);
-  mainElement.appendChild(dataWrapper);
-  mainElement.appendChild(simulationPanel);
-
-  return mainElement;
-}
-
 // Initialize app if DOM element exists
 const projectSimulator = document.getElementById('project-simulator');
 if (projectSimulator) {
@@ -465,13 +253,8 @@ if (projectSimulator) {
 export {
   updateElementText,
   renderTaskRowHistograms,
-  saveSvgAsImage,
   applyGraphSettings,
   resetGraphSettings,
-  createLogoElement,
   startSimulation,
-  createHeader,
-  createAdvancedSettings,
-  createSimulationPanel,
   setupUi,
 };
